Attack file logging handler to default logger
This is the correct logger to attach to. This is the logger that is used
for directly logging, as well as having uncaught exceptions etc logged
through it.

The `.errorhandler` service is ONLY intended for handling the display to
the browser or CLI of uncaught exceptions and PHP errors. It should not
be logged to directly and does not get errors that are logged through
the default logging service.